# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "3.0.0",
    "title": "Handwriter Express Application",
    "description": "Handwriter Express Application Routes"
  },
  "host": "handwritingbackend.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/logins/show": {
      "get": {
        "tags": [
          "Login Activity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get all User Logins",
            "schema": {
              "type": "object",
              "properties": {
                "user_logins": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "61c89118cbf275400d4ddc01"
                      },
                      "userId": {
                        "type": "string",
                        "example": "61c890bbcbf275400d4ddbf9"
                      },
                      "loggedOut": {
                        "type": "boolean",
                        "example": false
                      },
                      "loggedInAt": {
                        "type": "string",
                        "example": "2021-12-26T15:53:00.797Z"
                      },
                      "loggedOutAt": {
                        "type": "string",
                        "example": "2021-12-26T15:53:00.797Z"
                      },
                      "ipAddress": {
                        "type": "string",
                        "example": "::1"
                      },
                      "tokenId": {
                        "type": "string",
                        "example": "9109ZFMP1014EAAU"
                      },
                      "tokenSecret": {
                        "type": "string",
                        "example": "77777979ENSVHLIO75977759OFZNRMEM"
                      },
                      "tokenDeleted": {
                        "type": "boolean",
                        "example": false
                      },
                      "device": {
                        "type": "string",
                        "example": "PostmanRuntime/7.28.4"
                      },
                      "__v": {
                        "type": "number",
                        "example": 0
                      },
                      "current": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/logins/delete/{login_id}": {
      "get": {
        "tags": [
          "Login Activity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "login_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Delete selected user login from token ID",
            "schema": {
              "type": "object",
              "properties": {
                "deleted": {
                  "type": "boolean",
                  "example": true
                },
                "userLogin": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "example": "61c890bbcbf275400d4ddbfc"
                    },
                    "userId": {
                      "type": "string",
                      "example": "61c890bbcbf275400d4ddbf9"
                    },
                    "loggedOut": {
                      "type": "boolean",
                      "example": false
                    },
                    "loggedInAt": {
                      "type": "string",
                      "example": "2021-12-26T15:53:00.797Z"
                    },
                    "loggedOutAt": {
                      "type": "string",
                      "example": "2021-12-26T15:53:00.797Z"
                    },
                    "ipAddress": {
                      "type": "string",
                      "example": "::1"
                    },
                    "tokenId": {
                      "type": "string",
                      "example": "1114SEQI4644QYMY"
                    },
                    "tokenSecret": {
                      "type": "string",
                      "example": "27999797XVNSAHFV22272222GKZJUGSM"
                    },
                    "tokenDeleted": {
                      "type": "boolean",
                      "example": true
                    },
                    "device": {
                      "type": "string",
                      "example": "PostmanRuntime/7.28.4"
                    },
                    "__v": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/logins/delete/all/not-current": {
      "get": {
        "tags": [
          "Login Activity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Log out from all devices except current",
            "schema": {
              "type": "object",
              "properties": {
                "deleted": {
                  "type": "boolean",
                  "example": true
                },
                "userLogin": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "61c89118cbf275400d4ddc01"
                      },
                      "userId": {
                        "type": "string",
                        "example": "61c890bbcbf275400d4ddbf9"
                      },
                      "loggedOut": {
                        "type": "boolean",
                        "example": false
                      },
                      "loggedInAt": {
                        "type": "string",
                        "example": "2021-12-26T15:53:00.797Z"
                      },
                      "loggedOutAt": {
                        "type": "string",
                        "example": "2021-12-26T15:53:00.797Z"
                      },
                      "ipAddress": {
                        "type": "string",
                        "example": "::1"
                      },
                      "tokenId": {
                        "type": "string",
                        "example": "9109ZFMP1014EAAU"
                      },
                      "tokenSecret": {
                        "type": "string",
                        "example": "77777979ENSVHLIO75977759OFZNRMEM"
                      },
                      "tokenDeleted": {
                        "type": "boolean",
                        "example": false
                      },
                      "device": {
                        "type": "string",
                        "example": "PostmanRuntime/7.28.4"
                      },
                      "__v": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/user/logins/deletes/all": {
      "get": {
        "tags": [
          "Login Activity"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Logout from all devices.",
            "schema": {
              "type": "object",
              "properties": {
                "deleted": {
                  "type": "boolean",
                  "example": true
                },
                "user_login": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "example": "61c89118cbf275400d4ddc01"
                      },
                      "userId": {
                        "type": "string",
                        "example": "61c890bbcbf275400d4ddbf9"
                      },
                      "loggedOut": {
                        "type": "boolean",
                        "example": true
                      },
                      "loggedInAt": {
                        "type": "string",
                        "example": "2021-12-26T15:53:00.797Z"
                      },
                      "loggedOutAt": {
                        "type": "string",
                        "example": "2021-12-26T15:53:00.797Z"
                      },
                      "ipAddress": {
                        "type": "string",
                        "example": "::1"
                      },
                      "tokenId": {
                        "type": "string",
                        "example": "9109ZFMP1014EAAU"
                      },
                      "tokenSecret": {
                        "type": "string",
                        "example": "77777979ENSVHLIO75977759OFZNRMEM"
                      },
                      "tokenDeleted": {
                        "type": "boolean",
                        "example": true
                      },
                      "device": {
                        "type": "string",
                        "example": "PostmanRuntime/7.28.4"
                      },
                      "__v": {
                        "type": "number",
                        "example": 0
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Bad Request"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Token BlackListed"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Forbidden"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
...
  },
...
}